refactor(executor): decompose create_index.rs into focused modules
Breaks down the 1704-line create_index.rs file into smaller, focused modules
for better maintainability and testability:

New modules:
- validation.rs (~320 lines): Pre-creation validation including table existence,
  column validation, expression determinism checks, prefix length validation,
  and index name collision checks
- btree_index.rs (~116 lines): B-tree index creation for standard and unique indexes
- spatial_index.rs (~104 lines): R-tree/spatial index creation for geometry data
- vector_index.rs (~200 lines): IVFFlat and HNSW vector index creation
- expression_index.rs (~106 lines): Expression-based functional index creation

The main create_index.rs is now reduced to ~122 lines of dispatch logic,
with tests remaining in place for discoverability.

Closes #4868
